Set APPLICATION_EXTENSION_API_ONLY to YES on iOS target.
For many apps it makes sense to use AppAuth (OIDAuthState at least) in app extensions for example when loading new data from a today widget. However, AppAuth doesn’t enable the APPLICATION_EXTENSION_API_ONLY setting.

The only non-safe method call is to +[UIApplication sharedApplication]. With this change AppAuth checks whether the code is run from an app extension or not and then uses objc_msgSend() directly to access the sharedApplication without throwing a compiler error and returns an error when the framework user tried to present the authorization from an extension.