Allow user to specify dispatch queue for the `performActionWithFreshTokens` method
To date, all AppAuth blocks execute on the main queue. This allows the user to specify the queue for `performActionWithFreshTokens`. Unlike the rest of AppAuth's blocks, this one is called frequently – for every API call – thus it is important for the user to determine on which dispatch queue it executes.