Replaced UIApplication.keyWindow usage with searching for a fitting WindowScene
UIApplication.keyWindow was deprecated in iOS13 since there is no single keyWindow anymore when dealing with mutliple windows. Using keyWindow can result in the authenticationVC beign presented on a window that is not currently in the foreground. So the way to solve it is to look through all connected scenes and use a window of the most fitting windowScene.

Note:
When having multiple scenes of the same app in the foreground via splitView or slideover, this solution might result in the authenticationVC beign displayed on any of these scenes. But there is currently no better solution since there is no way to tell in which scene the user has triggered the authentication.